Refactored models in different files.
