+++ src/service/download_service.rs
@@ -1,5 +1,6 @@
 use crate::db::DB;
-use crate::models::itunes_models::{Podcast, PodcastEpisode};
+use crate::models::podcast_episode::PodcastEpisode;
+use crate::models::podcasts::Podcast;
 use crate::service::file_service::{FileService};
 use crate::service::mapping_service::MappingService;
