+++ src/service/path_service.rs
@@ -1,7 +1,8 @@
 use std::io;
 use std::path::Path;
 use crate::DbConnection;
-use crate::models::itunes_models::{Podcast, PodcastEpisode};
+use crate::models::podcast_episode::PodcastEpisode;
+use crate::models::podcasts::Podcast;
 
 use crate::service::file_service::{FileService, prepare_podcast_episode_title_to_directory};
